Bugfix - E2E test temp db discard SCRUM-107

* Add test environment configuration and update
logger to exclude test environment

* Add cross-env and cross-spawn packages for
cross-platform compatibility

* Refactor environment variable loading and database URI configuration.

* Fix env error handling in config.js


Approved-by: João R.

+++ drones-node-api/testRunner.mjs
@@ -1,51 +1,64 @@
-import { exec } from 'node:child_process';
+import spawn from 'cross-spawn';
 
 let childProcess = null;
 let sigintReceived = false;
 
-function execPromise(command) {
-  return new Promise((resolve, _reject) => {
-    childProcess = exec(command);
+function spawnPromise(script, envVars = {}) {
+  return new Promise((resolve, reject) => {
+    const env = { ...process.env, ...envVars };
+    const npmCmd = process.platform === 'win32' ? 'npm.cmd' : 'npm';
+    const args = ['run', script];
 
-    childProcess.stdout.on('data', (data) => {
-      console.log(data.toString());
-    });
+    // Use cross-spawn to handle the script execution
+    childProcess = spawn(npmCmd, args, { stdio: 'inherit', env });
 
-    childProcess.stderr.on('data', (data) => {
-      console.error(data.toString());
-    });
-
-    childProcess.on('close', (code) => {
+    childProcess.on('exit', (code) => {
       childProcess = null;
-      resolve();
+      if (code === 0) {
+        resolve();
+      } else {
+        const error = new Error(`Script "${script}" exited with code ${code}`);
+        error.code = code;
+        reject(error);
+      }
+    });
+
+    childProcess.on('error', (err) => {
+      childProcess = null;
+      reject(err);
     });
   });
 }
 
-async function runScript(script) {
+async function runScript(script, envVars = {}) {
   try {
-    await execPromise(`npm run ${script}`);
+    await spawnPromise(script, envVars);
   } catch (error) {
-    console.error(error.message);
+    console.error(`Script "${script}" run disturbed. `, error.message);
   }
 }
 
 async function handleSigint() {
-  console.log('\nShutting down and cleaning up due to SIGINT (Ctrl+C)');
+  console.log('\nShutting down and cleaning up due to SIGINT (Ctrl+C).');
   sigintReceived = true;
 
-  if (childProcess) {
+  if (childProcess && !childProcess.killed) {
     childProcess.kill('SIGINT');
   }
 
-  await runScript('e2eTest:physicalTeardown');
+  await runScript('e2eTest:physicalTeardown').catch((error) => {
+    console.error('Error during shutdown:', error.message);
+  });
+
+  process.exit();
 }
 
 process.on('SIGINT', handleSigint);
 
 (async () => {
   await runScript('e2eTest:setEnv');
-  await runScript('e2eTest:start');
+
+  await runScript('e2eTest:start', { NODE_ENV: 'test' });
 
   if (!sigintReceived) {
     await runScript('e2eTest:physicalTeardown');
